CalculatorクラスをCalcSvcYamamotoクラスへ移動
