+++ CalcLib/CalcLib.csproj
@@ -41,7 +41,6 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="CalcButton.cs" />
-    <Compile Include="Calculator.cs" />
     <Compile Include="CalcSvcMaeda.cs" />
     <Compile Include="CalcSvc.cs" />
     <Compile Include="CalcContext.cs" />
